// NextMigration/lib/api/types.ts
export interface CreateCard {
    term: string;
    definition: string;
    image?: string | null;
    number: number;
}

export interface UpdateCard {
    id: string;
    term?: string;
    definition?: string;
    number?: number;
    updated_at?: string;
}

export interface Card {
    id: string;
    term: string;
    definition: string;
    number: number;
    created_at: string;
    updated_at: string;
    set_id: string;
    owner_id: string;
}

export interface CardListResponse {
    cards: Card[];
}

// ============================================================
// SET LIKES
// ============================================================

export interface CreateSetLike {
    set_id: string;
}

export interface SetLike {
    id: string;
    user_id: string;
    set_id: string;
    set_type: 'studyset' | 'visualset';
    created_at: string;
}

export interface SetLikeListResponse {
    likes: SetLike[];
}

// ============================================================
// STUDYSETS
// ============================================================

export interface CreateStudyset {
    title: string;
    course: string;
    global_term_language: string;
    global_definition_language: string;
    folder_id: string;
    cardlist: CreateCard[];
}

export interface UpdateStudyset {
    title?: string;
    course?: string;
    global_term_language?: string;
    global_definition_language?: string;
    public_set?: boolean;
    cards?: UpdateCard[];
}

export interface Studyset {
    id: string;
    title: string;
    course: string;
    global_term_language: string;
    global_definition_language: string;
    created_at: string;
    last_updated: string;
    public_set: boolean;
    displayName: string;
    img_url: string;
    user_id: string;
    folder_id: string;
}

export interface FullStudyset extends Studyset {
    cards: Card[];
    likes: SetLike[];
    session?: StudySession;
    classrooms?: Classroom[];
    folders?: Folder[];
}

export interface StudysetListResponse {
    sets: Studyset[];
}

export interface AllSetsResponse {
    studysets: Studyset[];
    visualsets: Visualset[];
}
// ...
export interface Visualset {
    id: string;
    title: string;
    course: string;
    created_at: string;
    last_updated: string;
    public_set: boolean;
    user_id: string;
    displayName: string;
    img_url: string;
    folder_id: string;
}
// ...
export interface Folder {
    id: string;
    name: string;
    owner_id: string;
}
// ...
export interface Classroom {
    id: string;
    name: string;
    owner_id: string;
    type: string;
    created_at: string;
    verified: boolean;
}
// ...
export interface StudySession {
    id: string;
    started_at: string;
    duration_min: number;
    ended_at?: string;
    set_index: number;
    accuracy: number;
    average_response_time: number;
    longest_focus_streak: number;
    last_seen?: string;
    last_studied?: string;
    user_id: string;
    set_id: string;
    set_type: 'studyset' | 'visualset';
}